move external host tags payload to agentchecks provider
