+++ pkg/metadata/v5/payload_gohai.go
@@ -62,7 +62,6 @@ type Payload struct {
 	HostPayload
 	ResourcesPayload
 	// TODO: host-tags
-	ExternalHostPayload
 	GohaiPayload
 }
