+++ pkg/metadata/v5/v5_test.go
@@ -10,7 +10,6 @@ import (
 	"testing"
 
 	"github.com/DataDog/datadog-agent/pkg/collector/py"
-	"github.com/DataDog/datadog-agent/pkg/metadata/externalhost"
 	python "github.com/sbinet/go-python"
 	"github.com/stretchr/testify/assert"
 )
@@ -30,29 +29,4 @@ func TestMain(m *testing.M) {
 func TestGetPayload(t *testing.T) {
 	pl := GetPayload("testhostname")
 	assert.NotNil(t, pl)
-}
-
-func TestExternalHostTags(t *testing.T) {
-	host1 := "localhost"
-	host2 := "127.0.0.1"
-	sourceType := "vsphere"
-	tags1 := []string{"foo", "bar"}
-	tags2 := []string{"baz"}
-	eTags1 := externalhost.ExternalTags{sourceType: tags1}
-	eTags2 := externalhost.ExternalTags{sourceType: tags2}
-	externalhost.SetExternalTags(host1, sourceType, tags1)
-	externalhost.SetExternalTags(host2, sourceType, tags2)
-
-	pl := GetPayload("")
-	hpl := pl.ExternalHostPayload.Payload
-	assert.Len(t, hpl, 2)
-	for _, elem := range hpl {
-		if elem[0] == host1 {
-			assert.Equal(t, eTags1, elem[1])
-		} else if elem[0] == host2 {
-			assert.Equal(t, eTags2, elem[1])
-		} else {
-			assert.Fail(t, "Unexpected value for hostname: %s", elem[0])
-		}
-	}
 }
